% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.r
\name{acc_dt2stars}
\alias{acc_dt2stars}
\title{Transform result of r5r::accessibiliy or st_accessibility from data.table to stars}
\usage{
acc_dt2stars(dt, cutoffs, percentiles, nom = "accessibility")
}
\arguments{
\item{dt}{data.table of accessibility}

\item{cutoffs}{cutoffs of isochrones}

\item{percentiles}{percentiles of time travel}

\item{nom}{name of the opportunity which accessibility is evaluated}
}
\description{
Transform result of r5r::accessibiliy or st_accessibility from data.table to stars
}
