% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{idINS2stars}
\alias{idINS2stars}
\title{Transform a data.frame or a sf with an idINS to a stars object.
It doesn't take into account the geometry of the sf, but the idINS informations.
Assumes crs 3035}
\usage{
idINS2stars(x, crop = NULL, idINS = "idINS", default_res = 200)
}
\arguments{
\item{x}{a data.frame or sf object with a idINS column.}

\item{crop}{a polygon of interest to crop the stars result.}

\item{idINS}{name of the idINS column. Default to idINS.}

\item{default_res}{if resolution isn't inscribed in idINS. Default to 200.}
}
\description{
Transform a data.frame or a sf with an idINS to a stars object.
It doesn't take into account the geometry of the sf, but the idINS informations.
Assumes crs 3035
}
