% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{residINS}
\alias{residINS}
\title{Get the resolution of the data.frame or sf from the idINS column}
\usage{
residINS(x, idINS = "idINS")
}
\arguments{
\item{x}{a data.frame or sf object with a idINS column.}

\item{idINS}{name of the idINS column. default to idINS.}
}
\description{
Get the resolution of the data.frame or sf from the idINS column
}
