% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.r
\name{st_accessibility}
\alias{st_accessibility}
\title{Calculation of accessibility on a stars object with an opportunity as attribute}
\usage{
st_accessibility(
  data,
  opportunities_colname,
  departures_colname = NULL,
  scope = NULL,
  speed_max = 60,
  duration_max = 30,
  r5r_core,
  mode = "WALK",
  mode_egress = "WALK",
  departure_datetime = Sys.time(),
  time_window = 1L,
  percentiles = 50L,
  decay_function = "step",
  cutoffs = 30L,
  decay_value = 1,
  max_walk_dist = Inf,
  max_bike_dist = Inf,
  max_trip_duration = 120L,
  walk_speed = 3.6,
  bike_speed = 12,
  max_rides = 3,
  max_lts = 2,
  n_threads = Inf,
  verbose = TRUE
)
}
\arguments{
\item{data}{Stars object of points with an information on opportunities}

\item{opportunities_colname}{Name of the opportunity attribute}

\item{departures_colname}{Name of an attribute to filter by, if the value is NA (Optional)}

\item{scope}{Maximal distance between origine and destination points, in km. If is NULL, scope is calculated with speed_max and duration_max}

\item{speed_max}{In km/h. By default, 60}

\item{duration_max}{in min. By default, 30 min which means scope is by default 30 km}
}
\description{
Calculation of accessibility on a stars object with an opportunity as attribute
}
\section{remarks }{
 missing values on opportunity are excluded from any calculation, but not tiles with a zero value.
Transforming these values to NA before executing this function may be a good idea.
}

