% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility2.r
\name{st_accessibility2}
\alias{st_accessibility2}
\title{Calculation of accessibility on a stars object with an opportunity as attribute}
\usage{
st_accessibility2(
  data,
  opportunities_colname,
  departures_colname = NULL,
  scope = NULL,
  r5r_core,
  mode = "WALK",
  mode_egress = "WALK",
  departure_datetime = Sys.time(),
  time_window = 1L,
  percentiles = 50L,
  decay_function = "step",
  cutoffs = 30L,
  decay_value = 1,
  max_walk_dist = Inf,
  max_bike_dist = Inf,
  max_trip_duration = 120L,
  walk_speed = 3.6,
  bike_speed = 12,
  max_rides = 3,
  max_lts = 2,
  n_threads = Inf,
  verbose = FALSE,
  progress = TRUE
)
}
\arguments{
\item{data}{Stars object of points with an information on opportunities}

\item{opportunities_colname}{string. The column name in the `destinations`
input that tells the number of opportunities in each location.
Defaults to "opportunities".}

\item{departures_colname}{Name of an attribute to filter by, if the value is NA (Optional)}

\item{scope}{Max scope in km. Default to NULL, evaluated by max(cutoffs) * speed(mode).}

\item{r5r_core}{a rJava object to connect with R5 routing engine}

\item{mode}{string. Transport modes allowed for the trips. Defaults to
"WALK". See details for other options.}

\item{mode_egress}{string. Transport mode used after egress from public
transport. It can be either 'WALK', 'BICYCLE', or 'CAR'.
Defaults to "WALK".}

\item{departure_datetime}{POSIXct object. If working with public transport
networks, please check \code{calendar.txt} within
the GTFS file for valid dates. See details for
further information on how datetimes are parsed.}

\item{time_window}{numeric. Time window in minutes for which r5r will
calculate travel times departing each minute. When using
frequency-based GTFS files, 5 Monte Carlo simulations will
be run for each minute in the time window. See details for
further information.}

\item{percentiles}{numeric vector. Defaults to '50', returning the accessibility
value for the median travel time computed for a given
time_window. If a numeric vector is passed, for example
c(25, 50, 75), the function will return accessibility
estimates for each percentile, by travel time cutoff. Only
the first 5 cut points of the percentiles are considered.
For more details, see R5 documentation at
'https://docs.conveyal.com/analysis/methodology#accounting-for-variability'}

\item{decay_function}{string. Choice of one of the following decay functions:
'step', 'exponential', 'fixed_exponential', 'linear',
and 'logistic'. Defaults to 'step', which yields
cumulative opportunities accessibility metrics.
More info in `details`.}

\item{cutoffs}{numeric. Cutoff times in minutes for calculating cumulative
opportunities accessibility when using the 'step decay function'.
This parameter has different effects for each of the other decay
functions: it indicates the 'median' (or inflection point) of
the decay curves in the 'logistic' and 'linear' functions, and
the 'half-life' in the 'exponential' function. It has no effect
when using the 'fixed exponential' function.}

\item{decay_value}{numeric. Extra parameter to be passed to the selected
`decay_function`.}

\item{max_walk_dist}{numeric. Maximum walking distance (in meters) to access
and egress the transit network, or to make transfers
within the network. Defaults to no restrictions as long
as `max_trip_duration` is respected. The max distance is
considered separately for each leg (e.g. if you set
`max_walk_dist` to 1000, you could potentially walk up
to 1 km to reach transit, and up to _another_  1 km to
reach the destination after leaving transit). Obs: if you
want to set the maximum walking distance considering
walking-only trips you have to set the `max_trip_duration`
accordingly (e.g. to set a distance of 1 km assuming a
walking speed of 3.6 km/h you have to set `max_trip_duration = 1 / 3.6 * 60`).}

\item{max_bike_dist}{numeric. Maximum cycling distance (in meters) to access
and egress the transit network. Defaults to no
restrictions as long as `max_trip_duration` is respected.
The max distance is considered separately for each leg
(e.g. if you set `max_bike_dist` to 1000, you could
potentially cycle up to 1 km to reach transit, and up
to _another_ 1 km to reach the destination after leaving
transit). Obs: if you want to set the maximum cycling
distance considering cycling-only trips you have to set
the `max_trip_duration` accordingly (e.g. to set a
distance of 5 km assuming a cycling speed of 12 km/h you
have to set `max_trip_duration = 5 / 12 * 60`).}

\item{max_trip_duration}{numeric. Maximum trip duration in minutes. Defaults
to 120 minutes (2 hours).}

\item{walk_speed}{numeric. Average walk speed in km/h. Defaults to 3.6 km/h.}

\item{bike_speed}{numeric. Average cycling speed in km/h. Defaults to 12 km/h.}

\item{max_rides}{numeric. The max number of public transport rides allowed in
the same trip. Defaults to 3.}

\item{max_lts}{numeric (between 1 and 4). The maximum level of traffic stress
that cyclists will tolerate. A value of 1 means cyclists will
only travel through the quietest streets, while a value of 4
indicates cyclists can travel through any road. Defaults to 2.
See details for more information.}

\item{n_threads}{numeric. The number of threads to use in parallel computing.
Defaults to use all available threads (Inf).}

\item{verbose}{logical. `TRUE` to show detailed output messages (the default).}

\item{progress}{logical. `TRUE` to show a progress counter. Only works when
`verbose` is set to `FALSE`, so the progress counter does not
interfere with R5's output messages. Setting `progress` to `TRUE`
may impose a small penalty for computation efficiency, because
the progress counter must be synchronized among all active
threads.}

\item{origins, destinations}{a spatial sf POINT object with WGS84 CRS, or a
data.frame containing the columns 'id', 'lon',
'lat'.}
}
\value{
A stars objects
}
\description{
Calculation of accessibility on a stars object with an opportunity as attribute
}
\section{remarks }{
 missing values on opportunity are excluded from any calculation, but not tiles with a zero value.
Transforming these values to NA before executing this function may be a good idea.
}

