% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{stars2dt}
\alias{stars2dt}
\title{Transform a stars object to a data.table with idINS.
Assumes crs 3035}
\usage{
stars2dt(x, resolution = 200)
}
\arguments{
\item{x}{a stars object}

\item{resolution}{default to 200. Size of raster cells which are identified.}
}
\description{
Transform a stars object to a data.table with idINS.
Assumes crs 3035
}
